refactor(workflow): rename and restructure list filter components
- Renamed `FilterBy` to `FilterCondition` for clarity.
- Moved `models.py` to `entities.py` to better reflect the purpose of the contained classes.
- Updated imports and references across the codebase to align with the renaming and file restructuring.
